updates egoJwt and pages utils
