+++ client/global/utils/egoJwt/index.ts
@@ -12,3 +12,5 @@ export const canReadProgram = TokenUtils.canReadProgram;
 export const canWriteProgram = TokenUtils.canWriteProgram;
 export const isProgramAdmin = TokenUtils.isProgramAdmin;
 export const canReadSomeProgram = TokenUtils.canReadSomeProgram;
+export const canWriteSomeProgram = TokenUtils.canWriteSomeProgram;
+export const getReadableProgramShortNames = TokenUtils.getReadableProgramShortNames;
