+++ client/global/utils/pages.tsx
@@ -1,16 +1,39 @@
 import * as React from 'react';
-import { PROGRAMS_LIST_PATH } from 'global/constants/pages';
+import {
+  LOCAL_STORAGE_REDIRECT_KEY,
+  LOGIN_PAGE_PATH,
+  PROGRAMS_LIST_PATH,
+  DCC_OVERVIEW_PATH,
+  USER_PAGE_PATH,
+  PROGRAM_MANAGE_PATH,
+  PROGRAM_SHORT_NAME_PATH,
+  RDPC_PATH,
+  PROGRAM_DASHBOARD_PATH,
+} from 'global/constants/pages';
+import {
+  isDccMember,
+  isRdpcMember,
+  canReadSomeProgram,
+  getReadableProgramShortNames,
+} from './egoJwt';
 
-export const getRedirectPathForUser = (egoJwt: string) => {
-  /**
-   * TODO: actually implement this function
-   */
-  return PROGRAMS_LIST_PATH;
-  // if (isDccMember(egoJwt)) {
-  //   return DCC_OVERVIEW_PATH;
-  // } else {
-  //   return USER_PAGE_PATH;
-  // }
+export const getDefaultRedirectPathForUser = (
+  egoJwt: string,
+  useStatic: boolean = false,
+): string => {
+  if (isDccMember(egoJwt)) {
+    return PROGRAMS_LIST_PATH;
+  } else if (isRdpcMember(egoJwt)) {
+    return RDPC_PATH;
+  } else if (canReadSomeProgram(egoJwt)) {
+    const readableProgramShortNames = getReadableProgramShortNames(egoJwt);
+    const orderedProgramShortNames = orderBy(readableProgramShortNames);
+    return egoJwt
+      ? PROGRAM_DASHBOARD_PATH
+      : PROGRAM_DASHBOARD_PATH.replace(PROGRAM_SHORT_NAME_PATH, orderedProgramShortNames[0]);
+  } else {
+    return USER_PAGE_PATH;
+  }
 };
 
 export type GetInitialPropsContext = {
